2022/pyproject.toml
[tool.poetry]
name = "aoc2022"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.6.0"
Pillow = "^9.3.0"
numpy = "^1.23.5"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
black = "^22.10.0"
pytest = "^7.2.0"
types-Pillow = "^9.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
